fix(automation): identify 0148's own work by the ledger, not by the audit log
The restore excluded rows carrying an 'archive' audit entry, on the reasoning
that 0148 is raw SQL and writes none while a human archive always does. That
is true and insufficient: a row archived BEFORE 0148 by any path that did not
audit also carries no entry, and 0148 never touched it — it only matched
archived_at IS NULL. The restore was handing back rows 0148 never took.

dbmigrate runs a migration's SQL and its schema_migrations_core INSERT in one
transaction, and Postgres now() is that transaction's start time, so 0148's
archived_at and its ledger row's applied_at are the same instant to the
microsecond. Matching on it is exact provenance rather than an inference:
anything archived earlier is excluded by construction, and an absent ledger
row restores nothing, which is the right answer for a database 0148 never ran
on.

The suites now apply a migration the way dbmigrate does — SQL and ledger row
in one transaction — because the instant they share is the thing under test.

The new case fails against the audit-based predicate with exactly the row it
was meant to protect: archived a day before 0148, handed back anyway.
